requirements.txt
absl-py==2.0.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
cloudpickle==2.2.1
Farama-Notifications==0.0.4
filelock==3.12.4
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
gymnasium==0.29.1
idna==3.4
importlib-metadata==6.8.0
# install==1.3.5
Jinja2==3.1.2
joblib==1.3.2
Markdown==3.4.4
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
oauthlib==3.2.2
Pillow==10.0.1
protobuf==4.24.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.5.2
pyparsing==3.1.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
torch==2.0.1
torch-geometric==2.3.1
-f https://download.pytorch.org/whl/cpu/torch_stable.html
-f https://data.pyg.org/whl/torch-2.0.1+cpu.html
torch-scatter==2.1.1
torch-sparse==0.6.17
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.16
Werkzeug==2.3.7
zipp==3.17.0
